docs: address launch readiness review feedback
