+++ netlify/functions/__tests__/public-submissions.test.ts
@@ -60,7 +60,7 @@ import { captureScreenshot } from "../../../src/lib/services/screenshot";
 import { uploadScreenshot } from "../../../src/lib/services/cloudinary";
 import { extractMetadata } from "../../../src/lib/services/metadata";
 import { TEST_SUBMISSION_RATE_LIMIT_SECRET } from "../../../src/test/rate-limit-fixtures";
-import { createMockContext, getPgQuery } from "./test-utils";
+import { createMockContext, createMockDb, getPgQuery } from "./test-utils";
 
 const RATE_LIMIT_SECRET = TEST_SUBMISSION_RATE_LIMIT_SECRET;
 
@@ -69,33 +69,6 @@ interface SubmissionCreated {
   type: "tool" | "resource";
   status: "pending";
   message: string;
-}
-
-function createMockDb() {
-  const transactionDb = {
-    execute: vi.fn().mockResolvedValue({ rows: [] }),
-    insert: vi.fn().mockReturnThis(),
-    values: vi.fn().mockReturnThis(),
-    onConflictDoUpdate: vi.fn().mockReturnThis(),
-    onConflictDoNothing: vi.fn().mockReturnThis(),
-    returning: vi.fn(),
-  };
-  const mockDb = {
-    execute: vi.fn().mockResolvedValue({ rows: [{ attempt_count: 1 }] }),
-    insert: vi.fn().mockReturnThis(),
-    values: vi.fn().mockReturnThis(),
-    onConflictDoUpdate: vi.fn().mockReturnThis(),
-    onConflictDoNothing: vi.fn().mockReturnThis(),
-    returning: vi.fn(),
-    transaction: vi.fn(),
-    transactionDb,
-  };
-  mockDb.transaction.mockImplementation(
-    async (
-      callback: (transaction: ReturnType<typeof getDb>) => Promise<unknown>,
-    ) => await callback(transactionDb as unknown as ReturnType<typeof getDb>),
-  );
-  return mockDb;
 }
 
 function createSubmissionRequest(
